Refactored handler to be a provider.  Refactored NetJson bits to a provider.
