+++ test/RService.IO.Tests/MockServices.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Http;
-using RService.IO.Abstractions;
+﻿using RService.IO.Abstractions;
 
 namespace RService.IO.Tests
 {
